Pushing Tutorial Level
Its still super small but its filled in and I have some ideas on how I want it to play out and I think it aligns with what we discussed prior. I tried for a long time to get textures working but failed so I may need some help with that.